requirements.txt
PyQt6>=6.5.0
python-vlc>=3.0.18
Flask>=2.3.0
pytest>=7.0.0
mutagen>=1.47.0
